revert the space change in chapter 1
A survey conducted by Cardus reveals that the percentage of people in Canada who view the contribution of religion in daily life as “very bad” or “more bad than good” has risen sharply between 2017 to 2022.

“The rise of religious hate crimes is occurring against a backdrop of increasingly negative public attitudes toward the contributions of religion and faith communities to Canada,” the report said.

The report continued, “Majority opinions are not necessarily linked to hate crimes, of course — only a tiny fraction of Canadians commit hate crimes. Yet it is reasonable to assume that those who are accused of religious hate crimes may be emboldened by declining respect for or acceptance of religion and religious persons.”

Strong hostilities were expressed by atheists and agnostics. Almost half of the respondents believe that religious communities’ contribution in Canada is mostly negative.